[WIP] style: move LogbookForm styles to theme lib
